chore: cleanup result grid style
