+++ src/components/Search/SearchModal.tsx
@@ -1,8 +1,7 @@
 import clsx from 'clsx'
 import * as React from 'react'
 import Icon from 'components/Icon'
-import SearchItem from './SearchItem'
-import { SearchResult } from './index'
+import SearchItem, { SearchResult } from './SearchItem'
 
 export interface SearchModelProps {
   search: string
